+++ src/App.jsx
@@ -1,4 +1,3 @@
-
 import './App.css'
 
 import MeetingLobby from './pages/MeetingLobby'
@@ -6,66 +5,17 @@ import About from './pages/About'
 import MeetingDetails from './pages/MeetingDetails'
 import NotFoundPage from './pages/NotFoundPage'
 import Whiteboard from './pages/Whiteboard'
-import { NavLink ,useLocation, matchPath} from 'react-router-dom'
-
-import GIcon from './cmps/GIcon'
-
-
-
-
+import { useLocation, matchPath} from 'react-router-dom'
 
 import { Route, Routes } from 'react-router'
+import { AppHeader } from './cmps/AppHeader'
 
 function App() {
   console.log('App Rules Test');
-  
-
-  const routes = [
-    {
-      path: '/',
-      name: 'Home'
-    },
-
-    {
-      path: '/about',
-      name: 'About'
-    },
-    {
-      path: '/whiteboard',
-      name: 'Whiteboard'
-    }
-  ]
-
 
   return (
     <section className="app-container">
-      <header className="header bg-text bg-background-2 main-layout">
-        <div className="grid items-center grid-flow-col h-full">
-
-          <div className="logo">LOGO</div>
-          <nav className=''>
-            <ul className='grid grid-flow-col'>
-              {routes.map((route, index) => (
-                <li key={index}>
-                  <NavLink
-                    to={route.path}
-                    className={({ isActive }) => (isActive ? 'active-link' : '')}
-                  >
-                    {route.name}
-                  </NavLink>
-                </li>
-              ))}
-            </ul>
-          </nav>
-          <div className="actions">
-      
-          
-            <GIcon iconName='Home'/>
-         
-            <span>help</span> | <span>calender</span> | <span>user-preview</span>
-          </div>
-        </div>
-      </header>
+      <AppHeader />
 
       <main className="router main-layout">
         <Routes>
